+++ gusto_embedded/src/main/java/com/gusto/embedded_api/Employees.java
@@ -62,8 +62,8 @@ import com.gusto.embedded_api.utils.SerializedBody;
 import com.gusto.embedded_api.utils.Utils.JsonShape;
 import com.gusto.embedded_api.utils.Utils;
 import java.io.InputStream;
-import java.lang.Double;
 import java.lang.Exception;
+import java.lang.Long;
 import java.lang.Object;
 import java.lang.String;
 import java.net.http.HttpRequest;
@@ -1154,8 +1154,8 @@ public class Employees implements
      */
     public GetV1EmployeesEmployeeIdCustomFieldsResponse getCustomFields(
             String employeeId,
-            Optional<Double> page,
-            Optional<Double> per,
+            Optional<Long> page,
+            Optional<Long> per,
             Optional<? extends VersionHeader> xGustoAPIVersion) throws Exception {
         GetV1EmployeesEmployeeIdCustomFieldsRequest request =
             GetV1EmployeesEmployeeIdCustomFieldsRequest
