+++ gusto_embedded/src/main/java/com/gusto/embedded_api/Locations.java
@@ -37,8 +37,8 @@ import com.gusto.embedded_api.utils.SerializedBody;
 import com.gusto.embedded_api.utils.Utils.JsonShape;
 import com.gusto.embedded_api.utils.Utils;
 import java.io.InputStream;
-import java.lang.Double;
 import java.lang.Exception;
+import java.lang.Long;
 import java.lang.Object;
 import java.lang.String;
 import java.net.http.HttpRequest;
@@ -297,8 +297,8 @@ public class Locations implements
      */
     public GetV1CompaniesCompanyIdLocationsResponse get(
             String companyId,
-            Optional<Double> page,
-            Optional<Double> per,
+            Optional<Long> page,
+            Optional<Long> per,
             Optional<? extends VersionHeader> xGustoAPIVersion) throws Exception {
         GetV1CompaniesCompanyIdLocationsRequest request =
             GetV1CompaniesCompanyIdLocationsRequest
