+++ gusto_embedded/src/main/java/com/gusto/embedded_api/Payrolls.java
@@ -70,8 +70,8 @@ import com.gusto.embedded_api.utils.Utils.JsonShape;
 import com.gusto.embedded_api.utils.Utils;
 import java.io.InputStream;
 import java.lang.Boolean;
-import java.lang.Double;
 import java.lang.Exception;
+import java.lang.Long;
 import java.lang.Object;
 import java.lang.String;
 import java.net.http.HttpRequest;
@@ -495,8 +495,8 @@ public class Payrolls implements
      */
     public GetV1CompaniesCompanyIdPayrollReversalsResponse getApprovedReversals(
             String companyId,
-            Optional<Double> page,
-            Optional<Double> per,
+            Optional<Long> page,
+            Optional<Long> per,
             Optional<? extends VersionHeader> xGustoAPIVersion) throws Exception {
         GetV1CompaniesCompanyIdPayrollReversalsRequest request =
             GetV1CompaniesCompanyIdPayrollReversalsRequest
