+++ gusto_embedded/src/main/java/com/gusto/embedded_api/models/operations/GetInvoicesInvoicePeriodRequest.java
@@ -9,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.gusto.embedded_api.models.components.VersionHeader;
 import com.gusto.embedded_api.utils.SpeakeasyMetadata;
 import com.gusto.embedded_api.utils.Utils;
-import java.lang.Double;
+import java.lang.Long;
 import java.lang.Override;
 import java.lang.String;
 import java.lang.SuppressWarnings;
@@ -28,13 +28,13 @@ public class GetInvoicesInvoicePeriodRequest {
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
     @SpeakeasyMetadata("queryParam:style=form,explode=true,name=page")
-    private Optional<Double> page;
+    private Optional<Long> page;
 
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
     @SpeakeasyMetadata("queryParam:style=form,explode=true,name=per")
-    private Optional<Double> per;
+    private Optional<Long> per;
 
     /**
      * Filter companies returned in the active_companies response, will return an error if company not active during provided invoice period. i.e. `?company_uuids=781922d8-e780-4b6b-bf74-ee303166d022,bbbca930-7322-491c-ba7f-98707a52a9c5`
@@ -51,8 +51,8 @@ public class GetInvoicesInvoicePeriodRequest {
     @JsonCreator
     public GetInvoicesInvoicePeriodRequest(
             String invoicePeriod,
-            Optional<Double> page,
-            Optional<Double> per,
+            Optional<Long> page,
+            Optional<Long> per,
             Optional<String> companyUuids,
             Optional<? extends VersionHeader> xGustoAPIVersion) {
         Utils.checkNotNull(invoicePeriod, "invoicePeriod");
@@ -84,7 +84,7 @@ public class GetInvoicesInvoicePeriodRequest {
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
     @JsonIgnore
-    public Optional<Double> page() {
+    public Optional<Long> page() {
         return page;
     }
 
@@ -92,7 +92,7 @@ public class GetInvoicesInvoicePeriodRequest {
      * Number of objects per page. For majority of endpoints will default to 25
      */
     @JsonIgnore
-    public Optional<Double> per() {
+    public Optional<Long> per() {
         return per;
     }
 
@@ -129,7 +129,7 @@ public class GetInvoicesInvoicePeriodRequest {
     /**
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
-    public GetInvoicesInvoicePeriodRequest withPage(double page) {
+    public GetInvoicesInvoicePeriodRequest withPage(long page) {
         Utils.checkNotNull(page, "page");
         this.page = Optional.ofNullable(page);
         return this;
@@ -138,7 +138,7 @@ public class GetInvoicesInvoicePeriodRequest {
     /**
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
-    public GetInvoicesInvoicePeriodRequest withPage(Optional<Double> page) {
+    public GetInvoicesInvoicePeriodRequest withPage(Optional<Long> page) {
         Utils.checkNotNull(page, "page");
         this.page = page;
         return this;
@@ -147,7 +147,7 @@ public class GetInvoicesInvoicePeriodRequest {
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
-    public GetInvoicesInvoicePeriodRequest withPer(double per) {
+    public GetInvoicesInvoicePeriodRequest withPer(long per) {
         Utils.checkNotNull(per, "per");
         this.per = Optional.ofNullable(per);
         return this;
@@ -156,7 +156,7 @@ public class GetInvoicesInvoicePeriodRequest {
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
-    public GetInvoicesInvoicePeriodRequest withPer(Optional<Double> per) {
+    public GetInvoicesInvoicePeriodRequest withPer(Optional<Long> per) {
         Utils.checkNotNull(per, "per");
         this.per = per;
         return this;
@@ -239,9 +239,9 @@ public class GetInvoicesInvoicePeriodRequest {
  
         private String invoicePeriod;
  
-        private Optional<Double> page = Optional.empty();
+        private Optional<Long> page = Optional.empty();
  
-        private Optional<Double> per = Optional.empty();
+        private Optional<Long> per = Optional.empty();
  
         private Optional<String> companyUuids = Optional.empty();
  
@@ -263,7 +263,7 @@ public class GetInvoicesInvoicePeriodRequest {
         /**
          * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
          */
-        public Builder page(double page) {
+        public Builder page(long page) {
             Utils.checkNotNull(page, "page");
             this.page = Optional.ofNullable(page);
             return this;
@@ -272,7 +272,7 @@ public class GetInvoicesInvoicePeriodRequest {
         /**
          * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
          */
-        public Builder page(Optional<Double> page) {
+        public Builder page(Optional<Long> page) {
             Utils.checkNotNull(page, "page");
             this.page = page;
             return this;
@@ -281,7 +281,7 @@ public class GetInvoicesInvoicePeriodRequest {
         /**
          * Number of objects per page. For majority of endpoints will default to 25
          */
-        public Builder per(double per) {
+        public Builder per(long per) {
             Utils.checkNotNull(per, "per");
             this.per = Optional.ofNullable(per);
             return this;
@@ -290,7 +290,7 @@ public class GetInvoicesInvoicePeriodRequest {
         /**
          * Number of objects per page. For majority of endpoints will default to 25
          */
-        public Builder per(Optional<Double> per) {
+        public Builder per(Optional<Long> per) {
             Utils.checkNotNull(per, "per");
             this.per = per;
             return this;
