+++ gusto_embedded/src/main/java/com/gusto/embedded_api/models/operations/GetV1CompaniesCompanyIdContractorPaymentsRequest.java
@@ -10,7 +10,7 @@ import com.gusto.embedded_api.models.components.VersionHeader;
 import com.gusto.embedded_api.utils.SpeakeasyMetadata;
 import com.gusto.embedded_api.utils.Utils;
 import java.lang.Boolean;
-import java.lang.Double;
+import java.lang.Long;
 import java.lang.Override;
 import java.lang.String;
 import java.lang.SuppressWarnings;
@@ -53,13 +53,13 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
     @SpeakeasyMetadata("queryParam:style=form,explode=true,name=page")
-    private Optional<Double> page;
+    private Optional<Long> page;
 
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
     @SpeakeasyMetadata("queryParam:style=form,explode=true,name=per")
-    private Optional<Double> per;
+    private Optional<Long> per;
 
     /**
      * Determines the date-based API version associated with your API call. If none is provided, your application's [minimum API version](https://docs.gusto.com/embedded-payroll/docs/api-versioning#minimum-api-version) is used.
@@ -74,8 +74,8 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
             String endDate,
             Optional<String> contractorUuid,
             Optional<Boolean> groupByDate,
-            Optional<Double> page,
-            Optional<Double> per,
+            Optional<Long> page,
+            Optional<Long> per,
             Optional<? extends VersionHeader> xGustoAPIVersion) {
         Utils.checkNotNull(companyId, "companyId");
         Utils.checkNotNull(startDate, "startDate");
@@ -146,7 +146,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
     @JsonIgnore
-    public Optional<Double> page() {
+    public Optional<Long> page() {
         return page;
     }
 
@@ -154,7 +154,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
      * Number of objects per page. For majority of endpoints will default to 25
      */
     @JsonIgnore
-    public Optional<Double> per() {
+    public Optional<Long> per() {
         return per;
     }
 
@@ -237,7 +237,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
     /**
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
-    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPage(double page) {
+    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPage(long page) {
         Utils.checkNotNull(page, "page");
         this.page = Optional.ofNullable(page);
         return this;
@@ -246,7 +246,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
     /**
      * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
      */
-    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPage(Optional<Double> page) {
+    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPage(Optional<Long> page) {
         Utils.checkNotNull(page, "page");
         this.page = page;
         return this;
@@ -255,7 +255,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
-    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPer(double per) {
+    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPer(long per) {
         Utils.checkNotNull(per, "per");
         this.per = Optional.ofNullable(per);
         return this;
@@ -264,7 +264,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
     /**
      * Number of objects per page. For majority of endpoints will default to 25
      */
-    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPer(Optional<Double> per) {
+    public GetV1CompaniesCompanyIdContractorPaymentsRequest withPer(Optional<Long> per) {
         Utils.checkNotNull(per, "per");
         this.per = per;
         return this;
@@ -346,9 +346,9 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
  
         private Optional<Boolean> groupByDate = Optional.empty();
  
-        private Optional<Double> page = Optional.empty();
+        private Optional<Long> page = Optional.empty();
  
-        private Optional<Double> per = Optional.empty();
+        private Optional<Long> per = Optional.empty();
  
         private Optional<? extends VersionHeader> xGustoAPIVersion = Optional.empty();  
         
@@ -422,7 +422,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
         /**
          * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
          */
-        public Builder page(double page) {
+        public Builder page(long page) {
             Utils.checkNotNull(page, "page");
             this.page = Optional.ofNullable(page);
             return this;
@@ -431,7 +431,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
         /**
          * The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
          */
-        public Builder page(Optional<Double> page) {
+        public Builder page(Optional<Long> page) {
             Utils.checkNotNull(page, "page");
             this.page = page;
             return this;
@@ -440,7 +440,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
         /**
          * Number of objects per page. For majority of endpoints will default to 25
          */
-        public Builder per(double per) {
+        public Builder per(long per) {
             Utils.checkNotNull(per, "per");
             this.per = Optional.ofNullable(per);
             return this;
@@ -449,7 +449,7 @@ public class GetV1CompaniesCompanyIdContractorPaymentsRequest {
         /**
          * Number of objects per page. For majority of endpoints will default to 25
          */
-        public Builder per(Optional<Double> per) {
+        public Builder per(Optional<Long> per) {
             Utils.checkNotNull(per, "per");
             this.per = per;
             return this;
